test: Add comprehensive test suite with 152 tests (77% coverage)
Unit tests:
- DetectionOutput and BaseModelImpl (base classes)
- BaselineImpl and InvalidImpl (model implementations)
- DatasetLoader and download utilities
- Benchmark metrics (MSE, IoU, comparison, reporting)

Integration tests:
- FastAPI server endpoints (/health, /vram, /infer)
- Reference generation workflow

End-to-end tests:
- Complete inference cycle
- Output comparison between implementations
- Metrics collection and report generation

Fixes:
- Fix tensor dtype handling in compare_outputs for Long tensors

+++ src/benchmark/metrics.py
@@ -179,11 +179,11 @@ def compare_outputs(
         )
         return torch.cat([t, padding], dim=0)
     
-    # Pad tensors
-    pred_boxes_padded = pad_tensor(pred_boxes, max_len)
-    ref_boxes_padded = pad_tensor(ref_boxes, max_len)
-    pred_scores_padded = pad_tensor(pred_scores, max_len)
-    ref_scores_padded = pad_tensor(ref_scores, max_len)
+    # Pad tensors (ensure float for MSE calculation)
+    pred_boxes_padded = pad_tensor(pred_boxes.float(), max_len)
+    ref_boxes_padded = pad_tensor(ref_boxes.float(), max_len)
+    pred_scores_padded = pad_tensor(pred_scores.float(), max_len)
+    ref_scores_padded = pad_tensor(ref_scores.float(), max_len)
     pred_labels_padded = pad_tensor(pred_labels.float(), max_len).long()
     ref_labels_padded = pad_tensor(ref_labels.float(), max_len).long()
